// backend/src/types/generated/public/BotMessageTemplates.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.bot_message_templates */
export type BotMessageTemplatesEventType = string & { __brand: 'public.bot_message_templates' };

/** Represents the table public.bot_message_templates */
export default interface BotMessageTemplates {
  event_type: BotMessageTemplatesEventType;

  content: string;

  description: string | null;

  is_active: boolean | null;

  updated_at: string | null;
}

/** Represents the initializer for the table public.bot_message_templates */
export interface BotMessageTemplatesInitializer {
  event_type: BotMessageTemplatesEventType;

  content: string;

  description?: string | null;

  /** Default value: true */
  is_active?: boolean | null;

  /** Default value: now() */
  updated_at?: string | null;
}

/** Represents the mutator for the table public.bot_message_templates */
export interface BotMessageTemplatesMutator {
  event_type?: BotMessageTemplatesEventType;

  content?: string;

  description?: string | null;

  is_active?: boolean | null;

  updated_at?: string | null;
}